Add multiple groupings simulataneously

Add columns that bin rows into groups. Either use pax_add_groupings() to add any possible groupings, or individual pax_add_* functions to add a given group.

Usage

pax_add_groupings(
  tbl,
  groupings = pax_def_groupings(),
  ignore_missing_col = TRUE
)

pax_def_groupings(
  lgroups = seq(0, 200, 5),
  regions = list(all = 101:115),
  tgroup = NULL,
  ygroup = NULL,
  gear_group = NULL,
  ocean_depth = NULL
)

pax_add_lgroups(tbl, lgroups, ignore_missing_col = FALSE)

pax_add_regions(
  tbl,
  regions = list(all = 101:115),
  division_tbl = pax_temptbl(dbplyr::remote_con(tbl), "paxdat_gridcell"),
  ignore_missing_col = FALSE
)

pax_add_ocean_depth_class(
  tbl,
  ocean_depth_tbl = dplyr::tbl(dbplyr::remote_con(tbl), "ocean_depth"),
  breaks = c(0, 100, 200, 300),
  ignore_missing_col = FALSE,
  fill_resolution = 6
)

pax_add_gear_group(tbl, gear_group = NULL, ignore_missing_col = FALSE)

pax_add_temporal_grouping(
  tbl,
  tgroup = list(t1 = 1:6, t2 = 7:12),
  ignore_missing_col = FALSE
)

pax_add_yearly_grouping(tbl, ygroup = NULL, ignore_missing_col = FALSE)

pax_add_other()

Arguments

tbl

A dplyr query to apply grouping operations to

groupings

List of groupings generated by pax_def_groupings(), will apply each grouping operation in turn

ignore_missing_col

Boolean, should we ignore grouping columns that aren't present in the data?

lgroups

Vector of cut-off points. Values define lower bounds, everything above the final bound is considered part of the plus group

regions

List of region name to a vector of divisions, divisions are groups of gridcells described by division_tbl. Vectors can be pax_add_other() ("Other" value for gridcells without a matching division/region)

tgroup

List mapping temporal groups(seasons) to a vector of month integers, i.e. 1:12. Vectors can contain NA for any columns without a month value. If not supplied, list(year = 1:12) is used.

ygroup

List mapping year groups to a vector of year integers, i.e. 1:12. Vectors can contain NA for any columns without a year value. If not supplied, there will be one ygroup per year.

gear_group

List mapping gear names to a vector of mfdb_gear_code values. Vectors can contain NA (i.e. missing gear codes will be assigned here) or pax_add_other() ("Other" value used for unassigned gear codes)

ocean_depth, breaks

Ocean depth breaks, everything beyond the final group is part of a plus group

division_tbl

Mapping of gridcell -> division, by default the internal gridcell mapping

ocean_depth_tbl

Table containing a mapping from h3_cells to ocean_depth, used when the ocean_depth column is NA: the mean depth around the record's own h3_cells, see fill_resolution. Records without a position (or outside the table) stay NA, class "Unknown"

fill_resolution

H3 resolution at which the record's cells are matched to ocean_depth_tbl when filling missing depths (default 6, cells of about 36 km²; the bathymetry is too sparse for the database resolution). Cells without bathymetry at that resolution are matched at the two coarser resolutions

Value

pax_add_groupings

A dplyr query, with all grouping columns added

pax_def_groupings

List of grouping parameters to hand to pax_add_groupings()

pax_add_lgroups

Query with an additional lgroup column, binning a length column

pax_add_regions

Query with an additional region column, binning a gridcell column

pax_add_ocean_depth_class

Query with an additional ocean_depth_class column, binning a ocean_depth or h3_cells column

pax_add_gear_group

Query with an additional gear_name column, binning a mfdb_gear_code column

pax_add_temporal_grouping

Query with an additional tgroup column, binning a month column

pax_add_yearly_grouping

Query with an additional ygroup column, binning a year column

pax_add_other

Placeholder to mark "other" group